NERSystem/realtime_ner/server/requirements.txt
-i https://pypi.org/simple
alabaster==0.7.12
allennlp==0.8.1
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
aws-xray-sdk==0.95
awscli==1.16.93
babel==2.6.0
boto3==1.9.83
boto==2.49.0
botocore==1.12.83
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==7.0
colorama==0.3.9
conllu==0.11
cookies==2.2.1
cryptography==2.4.2
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
docker-pycreds==0.4.0
docker==3.7.0
docutils==0.14
ecdsa==0.13
editdistance==0.5.2
flaky==3.5.3
flask-cors==3.0.7
flask-sockets==0.2.1
flask==1.0.2
ftfy==5.5.0
future==0.17.1
gevent-websocket==0.10.1
gevent==1.3.6
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonnet==0.10.0 ; sys_platform != 'win32'
jsonpickle==1.0
kiwisolver==1.0.1
lark-parser==0.6.4
markupsafe==1.1.0
matplotlib==2.2.3
mock==2.0.0
more-itertools==5.0.0
moto==1.3.4
msgpack-numpy==0.4.3.2
msgpack==0.5.6
murmurhash==1.0.1
nltk==3.4
numpy==1.16.0
numpydoc==0.8.0
overrides==1.9
packaging==19.0
parsimonious==0.8.0
pbr==5.1.1
plac==0.9.6
pluggy==0.8.1
preshed==2.0.1
protobuf==3.6.1
py==1.7.0
pyaml==18.11.0
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.7.3
pygments==2.3.1
pyparsing==2.3.1
pytest==4.1.1
python-dateutil==2.7.5 ; python_version >= '2.7'
python-jose==2.0.2
pytorch-pretrained-bert==0.3.0
pytz==2017.3
pyyaml==3.13
regex==2018.1.10
requests==2.21.0
responses==0.10.5
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
spacy==2.0.18
sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
tensorboardx==1.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
tqdm==4.29.1
ujson==1.35
unidecode==1.0.23
urllib3==1.24.1 ; python_version >= '3.4'
wcwidth==0.1.7
websocket-client==0.54.0
werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
yaspin==0.14.0
